functions display, keyboard and animations moved for they own files
