+++ includes/animations.h
@@ -1,0 +1,6 @@
+#ifndef ANIMATIONS_H
+#define ANIMATIONS_H
+
+void animations();
+
+#endif
