+++ includes/display.h
@@ -1,10 +1,8 @@
-#ifndef KEYBOARD_H
-#define KEYBOARD_H
+#ifndef DISPLAY_H
+#define DISPLAY_H
 
 #include <GL/glew.h>   
 #include <GL/glut.h> 
-
-extern GLboolean fanSwitch;
 extern double eyeX; 
 extern double eyeY;
 extern double eyeZ;
@@ -12,8 +10,6 @@ extern double refX;
 extern double refY;
 extern double refZ;
 
-void keyboard( unsigned char key, int x, int y );
-void instructions();
-
+void display(void);
 
 #endif
