+++ makefile
@@ -4,7 +4,7 @@ CXXFLAGS = -Wall -Iincludes
 LIBS = -lGL -lGLU -lglut -lGLEW
 
 # Lista de arquivos fonte
-SRC = src/main.cpp src/keyboard.cpp src/simpleDraw.cpp $(wildcard src/objects/*.cpp)
+SRC = src/main.cpp src/keyboard.cpp src/simpleDraw.cpp src/animations.cpp src/display.cpp $(wildcard src/objects/*.cpp)
 
 # Nome do executável
 TARGET = bedroom
