+++ src/animations.cpp
@@ -1,0 +1,7 @@
+#include<animations.h>
+#include<bedroomObjects.h>
+
+void animations(){
+    animateClock();
+    animateFan();
+}
